streamlit/requirements.txt
streamlit
requests
pillow
pdf2image
markdown2
pynvml
